reorganize event participant list components, add event button
